+++ config/initializers/authority.rb
@@ -38,12 +38,14 @@ Authority.configure do |config|
   #
   # Defaults are as follows:
   #
-  # config.abilities =  {
-  #   :create => 'creatable',
-  #   :read   => 'readable',
-  #   :update => 'updatable',
-  #   :delete => 'deletable'
-  # }
+  config.abilities =  {
+    :create => 'creatable',
+    :read   => 'readable',
+    :update => 'updatable',
+    :delete => 'deletable',
+    :moderate => 'moderatable',
+    :vote_for => 'votable'
+  }
 
   # LOGGER
   # ======
@@ -57,7 +59,7 @@ Authority.configure do |config|
   #
   # Some possible settings:
   # config.logger = Rails.logger                     # Log with all your app's other messages
-  # config.logger = Logger.new('log/authority.log')  # Use this file
+  config.logger = Logger.new('log/authority.log')  # Use this file
   # config.logger = Logger.new('/dev/null')          # Don't log at all (on a Unix system)
 
 end