+++ db/migrate/20140325052615_create_answer_values.rb
@@ -1,8 +1,10 @@
 class CreateAnswerValues < ActiveRecord::Migration
   def change
     create_table :answer_values do |t|
-      t.integer :answer_id
-      t.integer :answer_option_id
+      t.references :answer
+      t.references :answer_template
+      t.references :answer_option
+
       t.decimal :numeric_value
       t.string :text_value
       t.datetime :time_value