+++ test/controllers/static_controller_test.rb
@@ -1,0 +1,11 @@
+require 'test_helper.rb'
+
+class StaticControllerTest < ActionController::TestCase
+
+  test "should get about" do
+    get :about
+    assert_response :success
+  end
+
+
+end