+++ test/controllers/surveys_controller_test.rb
@@ -1,8 +1,6 @@
 require 'test_helper.rb'
 
 class SurveysControllerTest < ActionController::TestCase
-
-
   test "User can start a survey" do
     login(users(:social))
 
@@ -23,30 +21,32 @@ class SurveysControllerTest < ActionController::TestCase
 
     get :ask_question, question_id: question_flows(:survey_1).first_question.id, answer_session_id: assigns(:answer_session).id
     assert_response :success
+    assert_not_nil assigns(:answer)
 
   end
 
-  test "User can answer question on survey" do
-    login(users(:social))
+  test "User can view grouped question on survey" do
+    login(users(:has_incomplete_survey))
 
-    get :start_survey, question_flow_id: question_flows(:survey_1).id
+    get :ask_question, question_id: questions(:q3b).id, answer_session_id: answer_sessions(:incomplete).id
 
-    assert assigns(:answer_session)
     assert_response :success
+    assert_not_nil assigns(:group)
+    assert_not_nil assigns(:questions)
+  end
 
-    post :process_answer, { 'question_id' => question_flows(:survey_1).first_question.id, 'answer_session_id' => assigns(:answer_session).id, question_flows(:survey_1).first_question.id.to_s => "blue", "direction" => "next"}
+  test "User can answer question on survey" do
+    login(users(:has_incomplete_survey))
+
+    post :process_answer, { 'question_id' => questions(:q3c).id, 'answer_session_id' => answer_sessions(:incomplete).id,  questions(:q3c).id.to_s => 22, "direction" => "next"}
 
     assert_redirected_to survey_report_path(assigns(:answer_session))
   end
 
   test "User can view survey report" do
-    login(users(:social))
+    login(users(:has_completed_survey))
 
-    get :start_survey, question_flow_id: question_flows(:survey_1).id
-    as_id = assigns(:answer_session).id
-    post :process_answer, { 'question_id' => question_flows(:survey_1).first_question.id, 'answer_session_id' => assigns(:answer_session).id, question_flows(:survey_1).first_question.id.to_s => "blue", "direction" => "next"}
-
-    get :show_report, answer_session_id: as_id
+    get :show_report, answer_session_id: answer_sessions(:complete)
 
     assert_response :success
   end