+++ test/models/vote_test.rb
@@ -1,0 +1,6 @@
+require "test_helper"
+
+class VoteTest < ActiveSupport::TestCase
+
+
+end